Mark uploaded binaries as compressed

+++ update.go
@@ -117,10 +117,9 @@ func updateCLI(channel string) {
 	if err := downloadBin(binPath+".new", build.URL); err != nil {
 		panic(err)
 	}
-	// FIXME: sha1 check doesn't work
-	//if fileSha1(binPath+".new") != build.Sha1 {
-	//	panic("SHA mismatch")
-	//}
+	if fileSha1(binPath+".new") != build.Sha1 {
+		panic("SHA mismatch")
+	}
 	os.Remove(binPath + ".old")
 	os.Rename(binPath, binPath+".old")
 	if err := os.Rename(binPath+".new", binPath); err != nil {
